+++ lib/feedjira/feed_utilities.rb
@@ -1,6 +1,8 @@
+# rubocop:disable Style/Documentation
+# rubocop:disable Style/DocumentationMethod
 module Feedjira
   module FeedUtilities
-    UPDATABLE_ATTRIBUTES = %w(title feed_url url last_modified etag)
+    UPDATABLE_ATTRIBUTES = %w(title feed_url url last_modified etag).freeze
 
     attr_writer   :new_entries, :updated, :last_modified
     attr_accessor :etag
@@ -11,8 +13,9 @@ module Feedjira
 
     module ClassMethods
       def parse(xml, &block)
+        xml = strip_whitespace(xml)
         xml = preprocess(xml) if preprocess_xml
-        super xml.lstrip, &block
+        super xml, &block
       end
 
       def preprocess(xml)
@@ -27,11 +30,20 @@ module Feedjira
       def preprocess_xml
         @preprocess_xml
       end
+
+      def strip_whitespace(xml)
+        if Feedjira.strip_whitespace
+          xml.strip
+        else
+          xml.lstrip
+        end
+      end
     end
 
     def last_modified
       @last_modified ||= begin
-        entry = entries.reject {|e| e.published.nil? }.sort_by { |entry| entry.published if entry.published }.last
+        published = entries.reject { |e| e.published.nil? }
+        entry = published.sort_by { |e| e.published if e.published }.last
         entry ? entry.published : nil
       end
     end
@@ -44,13 +56,13 @@ module Feedjira
       @new_entries ||= []
     end
 
-    def has_new_entries?
-      new_entries.size > 0
+    def new_entries?
+      !new_entries.empty?
     end
 
     def update_from_feed(feed)
       self.new_entries += find_new_entries_for(feed)
-      self.entries.unshift(*self.new_entries)
+      entries.unshift(*self.new_entries)
 
       @updated = false
 
@@ -60,7 +72,8 @@ module Feedjira
     end
 
     def update_attribute(feed, name)
-      old_value, new_value = send(name), feed.send(name)
+      old_value = send(name)
+      new_value = feed.send(name)
 
       if old_value != new_value
         send("#{name}=", new_value)
@@ -71,33 +84,36 @@ module Feedjira
     end
 
     def sanitize_entries!
-      entries.each {|entry| entry.sanitize!}
+      entries.each(&:sanitize!)
     end
 
     private
 
+    # This implementation is a hack, which is why it's so ugly. It's to get
+    # around the fact that not all feeds have a published date. However,
+    # they're always ordered with the newest one first. So we go through the
+    # entries just parsed and insert each one as a new entry until we get to
+    # one that has the same id as the the newest for the feed.
     def find_new_entries_for(feed)
-      # this implementation is a hack, which is why it's so ugly.
-      # it's to get around the fact that not all feeds have a published date.
-      # however, they're always ordered with the newest one first.
-      # So we go through the entries just parsed and insert each one as a new entry
-      # until we get to one that has the same id as the the newest for the feed
-      return feed.entries if self.entries.length == 0
-      latest_entry = self.entries.first
+      return feed.entries if entries.length.zero?
+
+      latest_entry = entries.first
       found_new_entries = []
+
       feed.entries.each do |entry|
-        if entry.entry_id.nil? && latest_entry.entry_id.nil?
-          break if entry.url == latest_entry.url
-        else
-          break if entry.entry_id == latest_entry.entry_id || entry.url == latest_entry.url
-        end
+        break unless new_entry?(entry, latest_entry)
         found_new_entries << entry
       end
+
       found_new_entries
     end
 
-    def existing_entry?(test_entry)
-      entries.any? { |entry| entry.id == test_entry.id }
+    def new_entry?(entry, latest)
+      nil_ids = entry.entry_id.nil? && latest.entry_id.nil?
+      new_id = entry.entry_id != latest.entry_id
+      new_url = entry.url != latest.url
+
+      (nil_ids || new_id) && new_url
     end
   end
 end